+++ internal/misc/status.go
@@ -1,4 +1,4 @@
-package telegram
+package misc
 
 type Status int
