+++ internal/telegram/room_number_status_handler.go
